CI: Remove update dependency job
Updating dependencies in a library does not make much sense.
Due to we dont have a pubspec.lock the app which uses this library always uses the
newest minor release by default.
For new major releases which have a new API this breaks anyway and needs manual help.
So if we depend on a specific major version we should not force the app to use the latest minor
versions without a reason to stay as compatible as possible.