CI: Clean up the CI and remove Moor tests
This cleans up the CI by removing the moor tests.
This allows the tests without olm to just use the google/dart image
like the code_analyze job. Also we just rename stages to test and deploy and
therefore rename the jobs a little bit.
This also removes the creation of the API docs
because they are now created by pub.dev anyway.