fix: Dont wait for first sync on init
This is a regression and was never intended. The init method should be
finished once the database is loaded completely.
return _sync(); was just to save one line of code. But then we made the init method
a Future<void> so we can now when the init method is terminated.
But this lead to the problem that it also waits for the first sync what we never wanted.