fix: Make type in AuthenticationData nullable
It seems that there is a problem in the matrix Spec according to this.
the Spec says that the type is required and session is nullable.
But for using this endpoint:
https://spec.matrix.org/unstable/client-server-api/#client-behaviour-1-1
We need to send an auth object which only contains the session.
We should file an issue at the spec and ask there for more information.
Currently we should make type nullable as well because otherwise
this would break uiaRequests in the app for SSO.