feat: Store sending attachment in temporary file storage
This could make retrying to send
files more easy as the file is
stored in the file storage.

This just replaces the current approach to store
the file in the memory instead so they are gone
on the next session.
For this we need a deleteFile() method for the
database.