feat: Add more SyncStatus updates for init
This makes it possible to track
what actually happens while
the Client runs the init() method.
The purpose is that clients can
give feedback to the user, what
the client is doing right now.
This is especially useful for
the login where we pipe the
Login, init and first sync method
behind each other.