refactor: Do not handle leave rooms on initial sync
This saves some ressources as
we do not need to delete rooms
as long as the database is empty anyway.
Also for some reason this fixes
an edge case where the keys
cache in the sqlite box gets
corrupted. It is unfortunate
that I wasn't able to find the
root cause or be able to reproduce
it. Wasted 4 hours now.