refactor: (BREAKING) Actually wipe ssss and cross signing when calling bootstrap wipe methods
Before we have just done nothing
when calling wipeCrossSigning() or
wipeSsss() in bootstrap and just
skipped to setup new. While
wiping usually means we want to re-init something, this
could lead to an undesired
state. Maybe the
consumer wants to just wipe
everything but start from scratch
later. I also think this could
lead to the case that some
bootstrap setups are
corrupted as we leave the
old recovery keys in
account data forever and
never wipe them.

The new behavior sends an
empty json object instead which
is the same behavior, Element
seem to do.